Improve the look of index and recipe pages
Added a class for the main image on each page to provide a consistent
look throughout the website.

Added a recipe-container and recipe class to improve the look of the
index page. The recipes are now displayed in a row, and each recipe is
displayed with a thumbnail and a link below it. Each recipe has a 3px
black border to separate it from the other recipes.

Padding, borders, and margins were adjusted on various elements to
improve the overall look of each page.